Only display locale resources on the home page
Currently, the Build with tBTC section pulls all resources. This means that
it's agnostic to locales, which means there can be an English home page that
lists non-English resources, even though there are enough English resources to
populate that space.

To fix this, the Resources component is removed, since it was a static
component. Instead, the home page's query now has two allMarkdownRemark
queries, distinguished via GraphQL aliases, to pull news items and resources.
The resources are rendered inline on the home-page template, very similarly to
how news items are.

One notable result is that in languages that _don't_ have enough resources,
the build section may be empty or short. This is an issue shared with news
items, so for now it'll have to do.