Add Prism for syntax highlighting for code snippets
To ensure Prism is loaded with only the languages, plugins, and settings
that we need (minimzing the number of things loaded into the build),
we have to add the babel-plugin, babel-plugin-prism js.
To use a custom .babelrc with Gatsby, we then have to add another preset
package, babel-preset-gatsby. Using the custom .babelrc causes a
comilation error with core-js modules, so we also are bumping Gatsby to
the latest version, which fixes that error.