Display Exchange List from GraphQl
Get data from Gatsby Netlify with proper query and display it on page as raw
<svg> element.

There were some problems with getting and coloring svg files when fetching them
from graphQl so I had to use react-svg library which allowed me to display raw
svg on the page (as <svg> semantic element).  In that case i could color the svg
by filling specific paths of it.