Try to fix filter invocation in sass
The version of node-sass we're using uses an old version of libsass,
which has an `invert` function as a SASS function, and conflicts with
the filter property's invert() function. Use interpolation to try and
get around this.

Note that bumping sass would fix this, but newer sass is incompatible
with the repo's versions of gatsby and related libraries, so we can't do
that :(